Create hooks to reduce the amount of logic on App
